Don't create multiple copies of attachments.
In some scenarios, if there were two different files with the same name that were ever sent by the same user, it could lead to a runaway behavior where the same file was copied every time the backup script run, resulting in many many extra copies.

This fixes that and also goes back and removes any of those excess files.